+++ graph/ExplicitCouplingGraph.h
@@ -11,28 +11,28 @@ using namespace std;
 
 class ExplicitCouplingGraph : public NormalizeCouplingWithChildren, public NormalizeSupportWithChildren, public NodeSetCouplingGraph {
 private:
-    vector<float> supports;
-    vector<vector<pair<uint, float>>> adj;
+    vector<double> supports;
+    vector<vector<pair<uint, double>>> adj;
     unordered_map<string, uint> node2i;
     vector<string> i2node;
 public:
     explicit ExplicitCouplingGraph(const string& name);
 
     const vector<string>& getNodeSet() override;
-    void add(const string& a, const string& b, float delta);
-    float get(const string& a, const string& b);
-    float getDirectCoupling(const string& a, const string& b) override;
+    void add(const string& a, const string& b, double delta);
+    double get(const string& a, const string& b);
+    double getDirectCoupling(const string& a, const string& b) override;
     vector<string> getDirectlyCoupled(const string& node);
 
-    void addSupport(const string& node, float delta);
-    float getSupport(const string& node);
-    float getAbsoluteSelfSupport(const string& node) override;
+    void addSupport(const string& node, double delta);
+    double getSupport(const string& node);
+    double getAbsoluteSelfSupport(const string& node) override;
 
-    void addAndSupport(const string& a, const string& b, float delta);
+    void addAndSupport(const string& a, const string& b, double delta);
 
-    void cutoffEdges(float minimumWeight);
-    void propagateDown(int layers = 1, float weightFactor = 0.2f);
-    void dilate(int iterations = 1, float weightFactor = 0.2f);
+    void cutoffEdges(double minimumWeight);
+    void propagateDown(int layers = 1, double weightFactor = 0.2f);
+    void dilate(int iterations = 1, double weightFactor = 0.2f);
 
 
     void plaintextSave(ostream& out) override;
@@ -40,12 +40,12 @@ public:
     void printStatistics() override;
 
     const unordered_set<string>& getChildren(const string &node) override;
-    float getNormalizedSupport(const string& node) override;
-    float getNormalizedCoupling(const string& a, const string& b) override;
+    double getNormalizedSupport(const string& node) override;
+    double getNormalizedCoupling(const string& a, const string& b) override;
 
 private:
     uint getNodeIndex(const string& node);
-    void addI(uint a, uint b, float delta);
+    void addI(uint a, uint b, double delta);
     unordered_map<uint, vector<uint>> getChildrenDict();
 };
