+++ graph/graphMixins/NormalizeCouplingWithChildren.cpp
@@ -44,19 +44,19 @@ void NormalizeCouplingWithChildren::getSelfAndDescendentsRec(const string& node,
     }
 }
 
-float NormalizeCouplingWithChildren::getDirectMultiCoupling(const string& a, const vector<string>& others) {
-    float result = 0;
+double NormalizeCouplingWithChildren::getDirectMultiCoupling(const string& a, const vector<string>& others) {
+    double result = 0;
     for (const auto& b : others) {
         result += getDirectCoupling(a, b);
     }
     return result;
 }
 
-float NormalizeCouplingWithChildren::getRelativeCoupling(const string& a, const string& b) {
+double NormalizeCouplingWithChildren::getRelativeCoupling(const string& a, const string& b) {
     return getRelativeMultiDirectCoupling(a, getSelfAndDescendants(b));
 }
 
-float NormalizeCouplingWithChildren::getRelativeMultiCoupling(const string& a, const vector<string>& others) {
+double NormalizeCouplingWithChildren::getRelativeMultiCoupling(const string& a, const vector<string>& others) {
     vector<string> directOthers;
     for (const auto& other: others) {
         getSelfAndDescendentsRec(other, directOthers);
@@ -64,36 +64,36 @@ float NormalizeCouplingWithChildren::getRelativeMultiCoupling(const string& a, c
     return getRelativeMultiDirectCoupling(a, directOthers);
 }
 
-float NormalizeCouplingWithChildren::getRelativeMultiDirectCoupling(const string& a, const vector<string>& others) {
+double NormalizeCouplingWithChildren::getRelativeMultiDirectCoupling(const string& a, const vector<string>& others) {
     if (others.empty()) {
         return 0;
     }
-    float result = getDirectMultiCoupling(a, others);
+    double result = getDirectMultiCoupling(a, others);
     for (const auto& child: getChildren(a)) {
         result += getRelativeMultiDirectCoupling(child, others);
     }
     return result;
 }
 
-float NormalizeCouplingWithChildren::getTotalRelativeCoupling(const string& a) {
+double NormalizeCouplingWithChildren::getTotalRelativeCoupling(const string& a) {
     auto cacheElement = total_relative_coupling_cache.find(a);
     if (cacheElement != total_relative_coupling_cache.end()) {
         return cacheElement->second;
     }
 
     vector<string> aCandidates = getCouplingCandidates(a, false);
-    float total_coupling = getRelativeMultiDirectCoupling(a, aCandidates);
+    double total_coupling = getRelativeMultiDirectCoupling(a, aCandidates);
     total_relative_coupling_cache[a] = total_coupling;
     return total_coupling;
 }
 
-float NormalizeCouplingWithChildren::getNormalizedCoupling(const string& a, const string& b) {
+double NormalizeCouplingWithChildren::getNormalizedCoupling(const string& a, const string& b) {
     const auto& nodeSet = getNodeSet();
     if (find(all(nodeSet), a) == nodeSet.end() || find(all(nodeSet), b) == nodeSet.end()) {
         return 0;
     }
-    float targetCoupling = getRelativeCoupling(a, b);
+    double targetCoupling = getRelativeCoupling(a, b);
     if (targetCoupling == 0) return 0;
-    float totalCoupling = getTotalRelativeCoupling(a);
+    double totalCoupling = getTotalRelativeCoupling(a);
     return targetCoupling / totalCoupling;
 }
